refactor(bff): align source layers with strict interface boundary
