+++ packages/bff/src/application/index.ts
@@ -1,6 +1,6 @@
-export * from "./orchestrator/aggregation.service";
-export * from "./orchestrator/mutation-orchestrator.service";
-export * from "./orchestrator/query-orchestrator.service";
-export * from "./orchestrator/query-pipeline.service";
+export * from "./orchestrator/aggregation.orchestrator";
+export * from "./orchestrator/mutation.orchestrator";
+export * from "./orchestrator/query.orchestrator";
+export * from "./orchestrator/query-pipeline.orchestrator";
+export * from "./services/meta-registry.service";
 export * from "./view/temporary-view-adapter.service";
-export * from "./view/view.contract";
