+++ packages/bff/src/bootstrap/app.module.ts
@@ -1,34 +1,34 @@
 import { randomUUID } from "node:crypto";
 import { Module } from "@nestjs/common";
 import { APP_FILTER } from "@nestjs/core";
-import { AggregationService } from "../../application/orchestrator/aggregation.service";
-import { CacheService } from "../../infra/cache/cache.service";
+import { AggregationService } from "../application/orchestrator/aggregation.orchestrator";
+import { MetaRegistryService } from "../application/services/meta-registry.service";
+import { TemporaryViewAdapter } from "../application/view/temporary-view-adapter.service";
+import { CacheService } from "../infra/cache/cache.service";
 import { AuditLogService } from "../common/audit-log.service";
 import { HttpExceptionFilter } from "../common/http-exception.filter";
-import { MetaController } from "../gateway/meta.controller";
-import { MetaRegistryService } from "../gateway/meta-registry.service";
-import { ViewController } from "../gateway/view.controller";
+import { MetaController } from "../interface/http/controller/meta.controller";
+import { ViewController } from "../interface/http/controller/view.controller";
 import {
   createRuntimeWsBroadcastBusFromEnv,
   parseRuntimeWsBroadcastBusMode,
   RUNTIME_WS_BROADCAST_BUS,
   RUNTIME_WS_INSTANCE_ID
-} from "../gateway/runtime-ws-broadcast.bus";
-import { RuntimeWsHealthController } from "../gateway/runtime-ws-health.controller";
-import { RuntimeWsOperationsState } from "../gateway/runtime-ws-operations.state";
+} from "../interface/ws/runtime-ws-broadcast.bus";
+import { RuntimeWsHealthController } from "../interface/ws/runtime-ws-health.controller";
+import { RuntimeWsOperationsState } from "../interface/ws/runtime-ws-operations.state";
 import {
   createRuntimeWsReplayStoreFromEnv,
   parseRuntimeWsReplayStoreMode,
   RUNTIME_WS_REPLAY_STORE
-} from "../gateway/runtime-ws-replay.store";
-import { QueryController } from "../gateway/query.controller";
-import { RuntimeWsGateway } from "../gateway/ws.gateway";
-import { AuditPersistenceService } from "../../infra/integration/audit-persistence.service";
-import { OrgScopeService } from "../../infra/integration/org-scope.service";
-import { PostgresQueryExecutorService } from "../../infra/integration/postgres-query-executor.service";
-import { MutationOrchestratorService } from "../../application/orchestrator/mutation-orchestrator.service";
-import { QueryOrchestratorService } from "../../application/orchestrator/query-orchestrator.service";
-import { TemporaryViewAdapter } from "../../application/view/temporary-view-adapter.service";
+} from "../interface/ws/runtime-ws-replay.store";
+import { QueryController } from "../interface/http/controller/query.controller";
+import { RuntimeWsGateway } from "../interface/ws/ws.gateway";
+import { AuditPersistenceService } from "../infra/integration/audit.service";
+import { OrgScopeService } from "../infra/integration/org-scope.service";
+import { PostgresQueryExecutorService } from "../infra/integration/postgres-query.service";
+import { MutationOrchestratorService } from "../application/orchestrator/mutation.orchestrator";
+import { QueryOrchestratorService } from "../application/orchestrator/query.orchestrator";
 
 @Module({
   imports: [],
