+++ packages/bff/src/bootstrap/index.ts
@@ -1,0 +1,5 @@
+export * from "./app.module";
+export * from "./bootstrap.service";
+export * from "./cli";
+export * from "./main";
+export * from "./migration-runner";
