+++ packages/bff/src/common/permission-errors.ts
