+++ packages/bff/src/index.ts
@@ -1,4 +1,6 @@
+export * from "./bootstrap";
 export * from "./application";
+export * from "./common";
 export * from "./domain";
 export * from "./infra";
 export * from "./interface";
