+++ packages/bff/src/infra/index.ts
@@ -1,4 +1,4 @@
 export * from "./cache/cache.service";
-export * from "./integration/audit-persistence.service";
+export * from "./integration/audit.service";
 export * from "./integration/org-scope.service";
-export * from "./integration/postgres-query-executor.service";
+export * from "./integration/postgres-query.service";
