+++ packages/bff/src/interface/contracts/meta-registry.contract.ts
