+++ packages/bff/src/interface/contracts/view.contract.ts
