+++ packages/bff/src/interface/index.ts
@@ -1,23 +1,12 @@
-export * from "./bootstrap/app.module";
-export * from "./bootstrap/bootstrap.service";
-export * from "./bootstrap/cli";
-export * from "./bootstrap/main";
-export * from "./bootstrap/migration-runner";
-export * from "./controller/app.controller";
-export * from "./common/audit-log.service";
-export * from "./common/http-exception.filter";
-export * from "./common/permission-errors";
-export * from "./common/request-id";
-export * from "./gateway/constants/meta-registry.constants";
-export * from "./gateway/contracts/meta.http";
-export * from "./gateway/contracts/meta-registry.contract";
-export * from "./gateway/meta-registry.service";
-export * from "./gateway/meta.controller";
-export * from "./gateway/query.controller";
-export * from "./gateway/view.controller";
-export * from "./gateway/view.http";
-export * from "./gateway/runtime-ws-broadcast.bus";
-export * from "./gateway/runtime-ws-health.controller";
-export * from "./gateway/runtime-ws-operations.state";
-export * from "./gateway/runtime-ws-replay.store";
-export * from "./gateway/ws.gateway";
+export * from "./contracts/meta-registry.contract";
+export * from "./contracts/view.contract";
+export * from "./protocols/meta.http";
+export * from "./protocols/view.http";
+export * from "./http/controller/meta.controller";
+export * from "./http/controller/query.controller";
+export * from "./http/controller/view.controller";
+export * from "./ws/runtime-ws-broadcast.bus";
+export * from "./ws/runtime-ws-health.controller";
+export * from "./ws/runtime-ws-operations.state";
+export * from "./ws/runtime-ws-replay.store";
+export * from "./ws/ws.gateway";
